debugged

+++ chk.sh
@@ -14,6 +14,10 @@ rm .done
 ##  Remove the "DONE" directory from the oldest modelDateTime directory before removing the whole directory.
 ##  This is to overcome api-dev delay in updating its list of models.
 modelDateTimes=(`s3cmd ls s3://modeltiles/HYCOM/ | cut -d/ -f5`)
+##  ALWAYS KEEP AT LEAST 3 MODEL DATETIMES
+if [[ ${#modelDateTimes[@]} -lt 3 ]]; then
+    exit
+fi
 oldest=${modelDateTimes[0]}
 s3cmd rm --recursive s3://modeltiles/HYCOM/${oldest}/DONE/
